Implement "is process graph supported" search
Implemented as "check if backend provides all processes and collections that this process graph contains".

I.e. this does *NOT* check whether
- the process graph is a valid openEO process graph,
- the backend would deem it valid (i.e. no query against the `/validation` backend),
- necessary variables are specified,
- variables have the appropriate types,
- variables have reasonable values (e.g. no check whether the requested band exists),
- the execution of the process graph would work without error,
- etc.

The implementation also depends on the "collection retrieval process" to be called `get_collection` and to specify the collection via a parameter called `name`. I'm not sure whether this is standardized or the earthengine-driver just happens to do it this way. Other backends not doing it this way will result in collection checking to be ignored, i.e. in such a case *only* the processes will be checked.