Add UI for the already existing search parameters
All search parameters that are already implemented (version, endpoints, collections, processes, process graph) can now be specified via the web app. The result is then requested from the API via Ajax and primitively displayed.

Changed the Ajax library from `request` to `axios` - I don't know what made me install `request` in the first place... I like `axios` and we use it for the js-client too, so better use that library.

The EndpointChooser could be generalised into a more versatile component, but for now it's good like it is.

It would be better if the constant would be dynamically derived from the openEO API spec, but I didn't yet come up with a smart way to specify the dependency.