Allow custom parameter names in endpoint paths
So far, `/collections/{collection_id}` wouldn't match when querying `/collections/{name}`, because the comparison was for strict string equality. Now a regex is used that allows anything but a closing curly bracket between the curly brackets, so backends can pick whatever they want as parameter names (which the API spec discourages, but permits).