Add appropriate metadata to backend search response
So far the backend just pumped out the entire `backendsWithCriteria` object without much consideration. Now it technically still does that, but the object gets restructured on the way there. It now includes the metadata to the requested collections and processes and has the `version` and `endpoint` fields on the root level so we can get rid of that `content` field that was just a database leftover.