Prettier displaying of search results
Instead of just dumping out the object the results are now output in a structured, prettier way.

The processes have the most elaborate rendering, realised by reusing components from the `processes-docgen` repo. Something similar should be developed for the collections. Outputting version and endpoints is rather unspectacular.

Note that the availability of DocGenProcesses depends on the changes in the `make-it-reusable` branch of my fork of the `processes-docgen` repo.