Use ProcessPanel directly, implement CollectionPanel
I like to have more control, so I rather put the loop together myself than relying on `DocGenProcesses` to do it, which may or may not result in lots of unnecessary other stuff being added in the same go.

I implemented a `CollectionPanel` by analogy to the existing `ProcessPanel` to display collections in a nicer way.

The new `start:dev` script allows to use the frontend while the backend is running and STILL use the Vue devtools (they are not available when Vue is compiled in production mode). On top of that, I think this script is a bit quicker.