Add map to BboxChooser, fix parsing of coordinates
Use a Leaflet map with OSM basemap to visualise the bbox. The bbox can be set by clicking on the map (first click = NW corner, second = SE). More convenient drawing and editing may be implemented later.
Switch from `parseInt` to `parseFloat`to fix truncation of coordinates when preparing params for the request.