Implement listing and searching for processes
Mostly like for collections, but with some differences:
- "summary" instead of "title"
- no extent
- add "exclude deprecated" switch
- add searching for parameter names and descriptions

Because the `parameters` object is a map, when searching for parameter names we don't search in an array but instead have to check for the existence of the appropriate keys.

To search through parameter descriptions we would need something like `process.parameters.*.description`. MongoDB doesn't support wildcards like that. Solution: Convert the `parameters` object to an array in the pipeline using the built-in function, then we can use `v`instead of the `*`.