Improve UI layout (tab-style navigation for sections)
Added a navigation in the style of tabs. The tabs toggle data variables, which are then again used in `v-show` directives. Moved the scrollbars from the `section` elements to the individual `panelContainer` divs so the tabs are sticky.
I hope those perfectly styled tabs look beautiful like this in all browsers, and not just in my Firefox...