Major refactoring: Split the huge Hub.vue into sub-components
But left the CSS styles centralised - they apply only to this specific use case of the components. Keeping the individual components plain vanilla allows for easy re-use.

Communication between the components is via events: The `XSearch` components `$emit` a `search-x` event with the search params as the argument. Event handlers are attached to the components in the main `Hub.vue` via `@search-x` directives. These forward the search params to the `queryX` handler methods, which then issue the actual requests (therefore we only need `axios` in the main file). The results are then stored as `matchedX` in the `Hub.vue` instance and forwarded to the `XResults` components via props to be displayed there.