Replace dependency `request` with `axios`
Switch from `request` to `axios` - the dependency was already removed in commit 6cd77c4, but I forgot to actually change the code... :see_no_evil: Didn't realise that yet because locally `request` was still installed and thus the `require` didn't fail.
Why change at all? The openeo-js-client uses axios, so the hub's frontend is using it too (although it doesn't use the js-client yet, but that may change in the future) - and when we already have that one library and it is also node.js compatible, why not use it for both the frontend and backend!