Improve UI: Collection and process search result rendering
CollectionResults and ProcessResults now use CollectionPanel and ProcessPanel (so far these were only included in the BackendResults).

CollectionPanel now outputs backend name and `retrieved` timestamp as special fields instead of hiding it in the general metadata. BackendResults also outputs the `retrieved` timestamp (and I built output of both properties into ProcessPanel in openeo-processes-docgen too).

In ProcessResults, added the conversion method that updates process documents that adhere to older openEO API specs to the latest spec, so the documents can be fed to ProcessPanel safely.

Add "emptyNotice" class because `em`s are now used for other things too, so a better selector was necessary.