Improve UI: Output extent of collections properly
So far, it was just spit out in the internal array format. Now the bounding box is output with North/South/East/West annotations and displayed on a Leaflet map. The map is not scrollable with the mouse wheel so that scrolling through the list is not interrupted by (usually unintended) map zooms. To avoid confusion, the map displayes the world only once (no wrap-around as is usually the case in web maps). The temporal extent is also printed in a somewhat nicer fashion, however a JSON-date-to-easily-human-readable converter should be added in the future.