Nicer "empty" notice, remove obsolete "Actions" sections
Before the first search, the results panel now tells the user how a search can be carried out. And when a search returned no results, there's now a nicer message.

The "Actions" section was a relict from when the "submit" buttons in the sticky headers wasn't there yet. It's now obsolete.