Add skeleton for Discover, Exchange and About sections
The search is just one modus of the Hub. The other two are discovering (aka just browsing through what's available) and exchanging/sharing process graphs so that others can use them (a bit like Gist). An About section doesn't hurt either.

Added the skeleton for that: Another `nav` in the header, the other sections, a `view` data field that stores which view we're currently displaying. Hiding of the inactive sections is done with a `.hidden` class instead of `v-show` because of the Leaflet map init problem (again).

Added basic content to the new sections:
- Discover: A list of all available backends (which so far really is just a list of URLs)
- Exchange: A "coming soon" notice
- About: A few words describing openEO Hub and links to the main openEO homepage + the hub's Github.

Adjusted the existing CSS so it all looks nice together.