Implement retrieving all backends including all their details
The response format should be identical with the one that is returned by the `POST /backends/search` endpoint so frontend components can be reused.

Added this to the `GET /backends` endpoint because a `GET` is more appropriate. Changing the representation of the resource depending on a query parameter (here: `details`) is absolutely fine.

Also added basic frontend output by sheepishly re-using the `Backend` component. Doesn't look very good yet and labels don't fit (e.g. "*matched* endpoints"), but it's something.

Getting those fields in the right structure was a bit of a challenge. I posted my conclusions here: https://stackoverflow.com/a/53732908/3746543