Improve EndpointChooser UI: Group endpoints into functionalities
The ordinary user doesn't (need to) know what each endpoint does, but is instead interested in the functionalities that are available.

The mapping of endpoints to functionality categories is a manual, somewhat subjective task. It was done together with @m-mohr.

Keeps the old functionality of EndpointChooser - if `categorizedEndpoints` prop is not given but `endpoints` is, it behaves like before.

I also didn't touch the old constant.

It would be possible to provide a toggle to switch between the two views, so that people who want to define their endpoint list more precisely could do so, but this use case is not common enough to justify the extra effort needed to code that.

This commit also removes a leftover `import` statement in the CollectionSearch component.