Implement very basic process graph repository (see #2)
Currently it's more or less a "pastebin" type of repository: You can paste your process graph, give it a title (ok, even a description!), submit it and it'll be online forever for everyone to see and you can't edit or delete it. Much additional work needed here, but as a first draft it's ok.