Deal with serviceTypes (or other fields) being non-existent or null
When the index is negative, MongoDB's `arrayElemAt` still returns an element, but counts from the back. Therefore, `serviceTypes` suddenly had the value `['collections', 'links']` in the frontend. This is now fixed on the backend side: The API now returns a proper `null` for the element if it is not present.
Now the frontend has to deal with elements being `null`, but that was accomplished quickly.