Move `normalizeProcess` calls out of Vue templates
Calling the normalization function from the template caused an issue with `parameters` not being set up properly in the process object because the function is invoked multiple times during rendering. So I moved it into a computed property. Actually, having the calls in a `prepare` function is much nicer anyways.

Also fixed `sortCallbackName` to work on either `id` or `name`, whichever is not `undefined`.