Fix `backends` collection storing ALL the raw data
The naming didn't really make sense: The collection called `backends` included all the dumped crawl data. This didn't make sense especially because there were collections called `collections` and `processes` that stored just those things that their names suggested; and because each of the three is an entity that can be searched for.

So now there's a new `raw` collection that holds all the crawled raw data, and the `backends` collection stores just the contents of the `GET_ALL_BACKENDS_PIPELINE`.

Nevertheless, the `backends` search is still executed against the `raw` collection -- changing that requires substantially more refactoring.

But at least the simple `GET /backends?details=true` request is now served via a simple "find all" operation instead of the complex aggregation.

I did a few functionality tests of the app to confirm that nothing got broken during the change.