Add UnsuccessfulCrawlNotice (aka "backend offline" flag)
Add notice to backends, collections and panels. But only when they are the top-level element, i.e. collections and processes within backends are not separately flagged.

Add the neccessary stuff to the pipelines.

Sort raw data by path so that the `/` entry is always first and thus the backend is assigned the `unsuccessfulCrawls` value of the `/` endpoint (the maximum doesn't make as much sense as I thought initially).

Also:
- Hub-specific extra information passed to the imported Process component via slots (e.g. DataRetrievedNotice) were only included in the ProcessResults component - now it is included in the Backend component too.
- Move `<div class="retrieved">` to DataRetrievedNotice component (in every case the component was used it was wrapped in such a `div`, so we can as well do that centrally).