Add config that allows to delay crawl requests
The Eurac backend once struggled with many requests in short succession (with that I mean: milliseconds after each other, if that). The error has since been fixed, but having a delay may be smart nevertheless. I set the default value to only 10ms because (1) delay is (currently) not a problem (anymore), and (2) otherwise crawling of the Earth Engine backend (400+ individual requests) takes ages (e.g. almost a minute when set to a mere 100ms), which is very annoying when debugging.