Replace imports from processes-docgen with new shared repo
So far, the `Process` component was imported from the `openeo-processes-docgen` repo, which was a fairly untidy thing to do. To improve that, @m-mohr meanwhile set up a shared `openeo-vue-components` repository that contains components upon which all openEO-related Vue.js web applications can build.

The normalization of processes (i.e. upgrading them to the latest spec) is now done by the `Process` component itself automatically, so no need anymore to import the `normalizeProcess` or `convertProcessToLatestSpec` functions.

When we don't need to normalize the processes, we don't need to prepare them at all anymore in the `ProcessResults` panel.

The `Process` component now includes its own CSS to style grouping, so this is no longer necessary in the Hub.

Note: Importing the components from the shared repo doesn't work with that repo's v0.3.2 just yet (the exports are wrong there), but this is due to be fixed in a release very soon).