Replace own Collection component with import from shared repo
Advanced displaying like the spatial extent map etc. is now done through slots.

Because this customization has quite a lot of code and is used twice, I wrapped it in a CollectionWrapper component which is then used in the Backend and CollectionResults components instead of the old Collection component.

All the code handling the collapsing etc. was moved to the shared repo, too.