Replace snippets in Backend comp with comps from shared repo
The new components SupportedFeatures, SupportedFileFormats, SupportedServiceTypes and BillingPlans replace the custom Vue snippets that so far took care of rendering lists in the template.

The SupportedFeatures component expects the endpoints in the format the openEO API uses, NOT in the `METHOD path` string format -> new section in `preparedBackend` that takes care of converting.

In return, the Hub no longer has to calculate the functionalities itself -> getting rid of the whole `functionalities` computed property and the `OPENEO_V0_3_1_FUNCTIONALITIES` constant import.

Since the calculation of functionalities is now no longer part of the Hub itself, getting the count of supported functionalities is a bit harder than before: We need to query the SupportedFeatures component for it. Because the `ref` to that component is only available when it's in the DOM, I had to change a `v-if` to a `v-show`. And because it's only available when it's already mounted (and not when the template is rendered for the first time), we can't call the function directly in the template but have to use a workaround with a `data` property that is populated once the `mounted` lifecycle hook is fired.