Dynamically retrieve backend titles from config (closes #18)
Instead of hard-coding the titles into the database during crawling, the titles are now dynamically retrieved from the config when a query is made.

It's done like this: Before actually sending the response, the root scope of the `data` object (or, if `data` is an array, the root scope of each array item) is checked for a `backend` property, and if it exists, it is replaced with `backendUrl` and `backendTitle` properties.

Another merit of this: The title is now included in *all* requests, not only in those for the `/` endpoint.

Also changed variable name in crawl.js from `backend` to `backendUrl` for better readability/clearity.