Nice error message when mongodb is not running (closes #11)
So far, the `assert` call raised an exception that wasn't handled anywhere and thus caused Node to display ugly error messages. Now the case that connecting to the db server fails is checked for and if it occurs, the MongoDB driver's error message is printed to the user. If the `--verbose` option is present, the user also gets to see the call stack.