Use common endpoints-to-functionalities map (closes #15)
As this mapping of endpoints to functionalities is needed in both the Hub and the `vue-components` repo, it has been moved to the `js-commons` repo in commit Open-EO/openeo-js-commons@6527369.

The common code has the endpoints (paths AND methods) in all-lowercase letters, so the search (in server.js) has to tolerate that.