Show warning sign if data is old or crawling failed (addresses #13)
Like @m-mohr suggested: "Put a warning sign after the title of the back-end and put the description [...] in a tooltip". The text is a bit more generic though: It doesn't show the crawl time, but just the statement that the data is older than the configured threshold. Likewise, it just states that the backend was recently unavailable for crawling, not how often this was the case or when the last successful crawl was.

This commit introduces a new config option `flagWhenOlderThanXHours`.