Major refactoring of how the server prepares and sends responses
Convention: All MongoDB wrappers return a Promise (except `getCollection`), i.e. the `.toArray()` is added to `find` and `aggregate` (it would be appended to any call of those two functions anyway, so we can just as well do it right there). This removes the need to check for `instanceof mongodb.Cursor` elsewhere (happened in `send` function so far).

Re-write `prepare` function: Now it not only takes care of adding the backend titles, but also removes the MongoDB IDs. More preparation handlers can easily be added by defining the function and adding it to the `callbacks` array. Prepare continues to accept both Promises and straight data.

The send function is no longer needed -- each endpoint handler now takes care of sending its result. It's better to have control of `next()` in the handler directly, and since the only other thing `send` did (extraction of data from a MongoDB Cursor object) is now obsolete, so is the `send` function itself. The handling of errors is now done by passing them to the `next` function, which then automatically creates an appropriate server error (status code 500 by default) and sends the error message.

Each of the changed endpoints was manually tested to ensure that it still works as before even after this major refactoring.

Also added some multiline comments to visually structure the code.