Lazy-load collections and processes in Backend comp (#12)
The `prepare` function on the backend side now accepts `additionalCallbacks` to run after the default preparation functions (removing MongoDB IDs and adding backend titles) have completed.

The `GET /backends` endpoint takes advantage of this by supplying a `clip` function that reduces the (usually very large) objects in the `collection` and `processes` arrays to just the `name` property, discarding all the details. This happens when the endpoint is called with `?details=clipped` appended.

To be able to lazy-load the full details later, a new endpoint `GET /backends/{backendUrl}` was added that returns the un-clipped version of the specified backend. The parameter has to be url-encoded (the serverside url-decoding happens automatically by Restify).

If the `details` parameter is anything else than `clipped`, the output is still the same as before.

The Backend component accepts a new `clippedDataSupplied` prop. If set to a truthy value it remembers that the supplied data is incomplete in the `dataComplete` data property. When the details are expanded for the first time, it fetches the full data from the server via an axios request, replaces the clipped data with it and remembers this by setting `dataComplete = true`.

To be able to take the data via a prop AND overwrite it via code, the `backend` prop had to be renamed to `backendData` (`backend` is now used only internally --
 it derives its value from the `backendData` prop by default, until it is overwritten by the `expand` function).

Warnings were added for configurations that I consider invalid:
1. Supplying clipped data but not collapsing initially (because the lazy-loading is coupled to the expanding mechanism and because users could click on things that are not completely loaded yet)
2. Supplying clipped data when the data is a search result (because the full data is loaded from the `/backends/{backendUrl}` endpoint which supplies the data in the default representation, not in the search representation.