Only show DataRetrievedNotice when data is old (#13)
The DataRetrievedNotice component was re-dedicated: So far it was *always* shown as an informative text, now it is displayed as a warning and only shown when the data is old.

Thus the restyling from "small, italic and grey" to "normal font size and style but red color".

In order to group similar information together, I moved all occurrences of `DataRetrievedNotice` next to the related `UnsuccessfulCrawlNotice`s.

And then I moved the two notices "beyond the break", i.e. they are only visible once the backend/collection/process has been expanded. This is still eye-catching enough because their presence is indicated by a warning sign next to the backend name, which is always visible.