Fix crawling of API v0.3 backends (removes js-client)
The newest version of the openeo-js-client only works with backends that support the openEO API v0.4. Backends with v0.3 can't be accessed because the client *requires* the `api_version` property in the capabilities, which wasn't part of the v0.3 spec.

Because the Hub only uses the js-client for a few rather simple things anyway, replacing those few things it does with a few lines of own code can restore v0.3 support. And although in the long run v0.3 won't be of too much relevance anymore, still supporting these older versions seems a nice-to-have feature for me "if we can have it this cheap".