Fix displaying of output formats and service types
The SupportedFileFormats and SupportedServiceTypes components expect the data as objects, not arrays. So simply don't convert to arrays.

But then we can't sort anymore
-> code removed
-> wrote issue for implementing this in the components themselves: Open-EO/openeo-vue-components#9

Also, the CSS needs to be NOT scoped to have effect on the child components. The `output-formats` are now called `file-formats`. The `ul.functionalities` rule is not needed anymore since the rendering of the functionalities was "outsourced".