Implement grouping of backends in frontend
Increased the vue-components version to 0.4.0 which will soon have the `Tab` and `Tabs` components which are needed for this. Then increased the Vue version to 2.6.0 because the vue-components now need this version.

Added `details=grouped` variant to the `/backends` endpoint which outputs the backends in groups as defined by the config. The returned backends are still clipped.

New component `BackendGroup` which wraps all backends of a group in tabs, outputs the group name and is collapsible (that really is all it does).

In the `Hub` component: Use `BackendGroup`s instead of individual `Backend`s - everything stays pretty much the same (just use `details=grouped` instead of `clipped`).

Only the sorting is now a bit different: First sort the groups by name, then the backends within the groups by version (down to the patch version, instead of the title as the last criterion) because the title is (at least semantically) identical for all backends within a group.