Fine-tune styling of backends in Discover section
Use "pills" style for Tabs in BackendGroup.

Nicer margins and padding

Introduced new props for `Backend`: `collapsible` and `showVersion`. New behaviour that can be achieved when setting them to false:
collapsible: Backend content can't be collapsed, the heading doesn't have `cursor: pointer` and the triangle.
showVersion: Hides the ` | v1.2.3` suffix. The prop is passed to the `BackendName` component which actually handles the behaviour.

In BackendGroup, both are set to false.