Implement filtering of backend groups by output format
New external component: Vue-Multiselect (see https://vue-multiselect.js.org/). It requires a stylesheet - which is included at the bottom of the DiscoverSection.vue file.

New API endpoint: Getting all output formats that are present in the database, sorted by count (so that more common formats appear further at the top of the autosuggest list). The MongoDB aggregation pipeline for this is in the dbqueries.js file.

First content for the new Filters section: The output format selector, created with the new VueMultiselect component.

All filter configurations will be saved into the`filters` object of the `DiscoverySection`.

`checkFilters` takes care of the filtering itself - for now locally in the browser, for more complex things in the future the function may issue async calls to the API.

Added some more CSS to make the Filters section look nicer.

Note that for now, the vertical separator lines between the BackendGroups don't work correctly when some of them are hidden via `v-show` due to the filtering.