Fix crawling of v0.4.x output formats (#17, point 2 of #26)
In openEO API v0.3.x, the actual output formats were to be found within the `formats` property of the return document of a backend's `/output_formats` endpoint.
In openEO API v0.4.x, this was changed -- the actual formats are now within the root element of the backend's reply.
Thus, the processing pipeline that prepares the data (which is executed after the actual crawling) must not access `.formats` any longer; instead, the entire `outputFormats` object is now simply passed on (that is what the `1` means).
The difference between the versions is to be handled by the frontend via the migrations in openeo-js-commons. The migrations do their work just fine, just the presentation has a minor bug, see Open-EO/openeo-vue-components#10 -- but that only applies to v0.3.x backends. Once that's fixed and released, the Hub will gain bug-free presentation of output formats of all backends by increasing the `vue-components` version number accordingly.