Implement "search by API v0.4 process graph" (#17)
I thought "It can't be *that* much work"...

The solution became a bit hacky though :angel: O:-)

We're hijacking the ProcessGraph class's validation procedure to extract processes and collections while the process graph is being traversed anyway.

Added a switch in the frontend to choose between the old and the new code -> new Boolean parameter `legacyProcessGraph`.

Co-Authored-By: Matthias Mohr <m.mohr@uni-muenster.de>