Implemented "search by any process graph" much simpler (#17)
And it's not hacky at all!

The old algorithm's dead-simple recursive parsing works quite well for v0.4 too -- but I only discovered that after the previous commit's code was already completed... The only thing that had to be adapted was the extraction of the collection names, because they are now called "id"s and the corresponding process was renamed from `get_collection` to `load_collection`.

But that's pretty much it, and the user doesn't even need to specify which version is being supplied! :tada: