Add capabilities document for backend API
Adding this should facilitate the usage of the `/process_graph` endpoints as if the Hub was a normal openEO backend.

The `api_version` is the openEO API version that the `/process_graphs` endpoints comply to, the backend's version and description are taken from the respective property in the `package.json`, the title shouldn't ever change, and the `endpoints` array is constructed manually so that only those endpoints are included that are actually intended (and fit for) public use. For now, these endpoints are the `/process_graph` endpoints and the configured list of backends under `/backends`.