Implement filtering by "backend has a free plan" (#34)
Actually the filtering is by "NOT every plan of every backend in the group is explicitly paid". This way, backends that don't supply billing information (maybe because there *is* no billing, i.e. everything's free) are not falsly excluded. This *should* work, because backends that are paid *must* include this information.