Handle "no filters" correctly when backend info is incomplete
Locally I usually only crawl GEE, R and VITO (because they can be crawled quickly), so when testing the new filtering on a database with *all* backends crawled, suddenly backends were missing in the default view! That was caused by the code returning `false` if the `outputFormats` or `serviceTypes` property wasn't present in a backend object. But that should of course not be the case when there are no filters present.
-> Added a check for empty filters for *all* array-based filters.