Implement filtering by collection identifiers (#34)
It's questionable whether that makes much sense as long as IDs are not unified across backends, because (probably?) any selection in this component reduces the filtering result to exactly 1 match. But a user may use this input field to look for e.g. DEM datasets and can then explore backends that have this kind of data by varying the choice of the particular dataset within this search input.

The search term of the auto-suggest is matched against both the identifier and the `title` attributes of the collections.

The search is carried not carried out by the Multiselect component because its search engine is rather limited (can't search for multiple terms separated by spaces and can't search within anything but the label).

Both `id` and `name` (i.e. v0.3 and v0.4) are supported as identifiers.

Uses a custom template for the `option` slot to display not only the label (the identifier) but also the `title` so the user has more of an idea what is behind the sometimes mysterious name.