Implement filtering by collection search terms (#34)
So far it was only possible to filter backends by collection *identifiers*. Now it's also possible to input search terms and the system filters out backends where none of the search terms is found anywhere in that backend's collections's identifiers or titles. So now one keyword in the "Collections" filter input (Multiselect) may match many collections at once!

I take advantage of the fact that when the "tag" (that's what it's called in Vue-Multiselect speech) is created, the result of searching for this term within the collections is known anyway (needed for the auto-complete suggestions). So I save this result in the tag's `matches` attribute and use this information again in the `checkFilters` method.

In the process of developing this, I got rid of all the references to the old `name` attribute (which is now `id`). Hence, I also dropped the `customLabel`.

Note: The removal of the references to the old `name` attribute *depends on changes in the backend that are not yet commited* (because it's not yet tested that these changes don't break any other parts of the code).