Replace all collection/process "name"s with "id"s
In openEO API v0.3.x, the identifiers for collections and processes were called "name", and this nomenclature still existed here and there in the Hub - even though they're now (from API v0.4.0) called "id". So now I finally purged the old term from the Hub's API (and subsequently components that use that API) by replacing the appropriate occurences of "name" with "id". (This is the necessary change that I mentioned in the previous commit 2b44839.) This change shouldn't break anything on the Hub, and as far as I know nobody else uses the API parts that are affected, so nothing should break anywhere else either.