Fix #37 (auto-complete inputs initially showing "List is empty")
The inputs take their suggestions from the optionCollections / optionProcesses (short: optionX) computed properties, which in turn just concatenate the searchedX and taggedX arrays, both of which are (of course) initially empty. The searchedX arrays are only populated by the searchX methods, so these have to be called once after the data has been fetched (with the empty string as the search keyword) to trigger the initial setup.