Group and sort auto-complete suggestions for coll./proc. (#38)
So far *all* collections and processes were returned, and in whatever order they were retrieved from the database. Now items with the same ID are grouped, and then sorted by size of the group (and after that alphabetically).

For processes, a new field "allSummaries" is returned so that the filter search can take all the different summaries of each backend into account. This could be done for collections too, but it's not really necessary there because usually there is only one instance of each collection anyway, so it would just be repetition. The process summaries, in contrast, are rather diverse, so it's necessary there.