Always order collection suggestions alphabetically (#38)
Instead of by count, because the count is always 1 or 2 anyway (depending on whether the backend is listed with 1 or 2 versions). And then the count doesn't need to be displayed anymore either.

The new ordering revealed that there was a collection with id=null and title=null, which I traced back to the old Mundialis backend. It returned an error instead of a valid collections object in its `GET /collections` response. So I introduced a safety check in the pipelines that populate the `collections` and `processes` MongoDB collections.