Remove BackendName component
It's no longer needed in the CollectionWrapper and ProcessWrapper components (it *was* needed there for displaying backend instances in the Search section, but that Search section is gone - and, thus, so is that need (makes sense)).

So the only spot where BackendName would still be needed is the Backend component itself. And well, *then* that logic can quite as well be implemented directly there, without the need for a sub-component.

Plus: The BackendName component *used* to be responsible for generating the warning sign. But that generating code must soon be moved to the BackendGroup component anyway (so that the icon can reside in the Tab title, see #49).

=> The BackendName component is obsolete and can be removed.

Also some minor minor change: Wrapping the backendUrl into a <p> paragraph.