Move warning sign icon to Tab title (fixes #49) and edit notices
When you look at this commit only, it's not a move but an implementation. But when you also look at the previous commit (16b6e5c), it's clear that it's actually a move.

Also add that icon to the texts of the DataRetrievedNotice and UnsuccesfulCrawlNotice so that there is a direct visual connection between the icon at the top and the explanatory text further down.

While I'm at it, I also removed those two notices from the ProcessWrapper because the option to display them there is not used anyway (not currently, and most probably not anytime soon either).
I did *not* remove them from the CollectionWrapper because collection details are not exclusively gathered from the `/` endpoint, but also crawled from the separate `/collections/{id}` endpoint, which may have a different "data freshness" status than the main part of the data (although the backend API, as of now, doesn't deliver that information).