Remove `parametersAsArray` field from "process" MongoDB collection
It's no longer needed because parameter names and descriptions are not searched anymore (the feature was removed when removing the Search section (it never was a really helpful feature anyway)).

Hence, the indexes are no longer needed either.

Running the new script against a MongoDB that still has the old index will result in an error - but I removed it manually from my local db, the deployed Hub drops the database on each restart and that's all the instances I know of, so that shouldn't be a problem.