Expect that all configured backends support well-known discovery
This is a BREAKING CHANGE because old configs may not be understood anymore.

Both of the other config options -- (1) supplying a direct URL to a backend and (2) supplying an object with _multiple_ direct URLs to backends -- are not supported anymore. I adapted the README accordingly.

All backends that currently exist (at least that we know of) currently DO support well-known discovery (excluding the R server, but it only supports openEO API v0.3.x anyway so it is no longer relevant).

If you look at a diff that ignores whitespace the change in crawl.js doesn't look nearly as big anymore.